+++ ModFolder/Loenn/entities/SeamlessSpinner.lua
@@ -46,6 +46,10 @@ seamlessSpinner.depth = function(room,entity) return entity.depth or -8500 end
 seamlessSpinner.fieldInformation = {
     foreDirectory = vivUtils.GetFilePathWithNoTrailingNumbers(false),
     backDirectory = vivUtils.GetFilePathWithNoTrailingNumbers(false),
+    coldCoreModeSpritePath = vivUtils.GetFilePathWithNoTrailingNumbers(false),
+    hotCoreModeSpritePath = vivUtils.GetFilePathWithNoTrailingNumbers(false),
+    coldCoreModeTriggerSpritePath = vivUtils.GetFilePathWithNoTrailingNumbers(false),
+    hotCoreModeTriggerSpritePath = vivUtils.GetFilePathWithNoTrailingNumbers(false),
     hitboxType={
         options={
             "default",
@@ -111,6 +115,7 @@ seamlessSpinner.ignoredFields = function(entity)
     local attrs = { "x", "y", "_id", "_name"}
     local non_dyn = {"solidTileCutoff", "dust"}
     local dyn = {"fgAnimDelay", "bgAnimDelay","fgAnimRandomize", "bgAnimRandomize", "fgFlipX", "fgFlipY", "bgFlipX", "bgFlipY"}
+    local core = {"coldCoreModeSpritePath", "hotCoreModeSpritePath", "coldCoreModeTriggerSpritePath", "hotCoreModeTriggerSpritePath"}
 
     if entity.dynamic then
         for _, item in ipairs (non_dyn) do
@@ -118,6 +123,12 @@ seamlessSpinner.ignoredFields = function(entity)
         end
     else
         for _, item in ipairs (dyn) do
+            table.insert(attrs, item)
+        end
+    end
+
+    if not entity.useCoreModeStyle then
+        for _, item in ipairs (core) do
             table.insert(attrs, item)
         end
     end